+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- [#4](https://github.com/MalteHerrmann/evmos-checker/pull/4) Add basic cobra command structure
+
 ### Improvements
 
 - [#7](https://github.com/MalteHerrmann/evmos-checker/pull/7) Add linters and Makefile
-